Implements transfer modal

+++ Domain/Dtos/TransferDto.cs
@@ -2,6 +2,6 @@ namespace Domain.Dtos {
     public class TransferDto {
         public int AccountFrom { get; set; }
         public int AccountTo { get; set; }
-        public double Amount { get; set; }
+        public float Amount { get; set; }
     }
 }
